% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meDCR.R
\name{CRC_test}
\alias{CRC_test}
\title{Distinguish between CRC and normal samples}
\usage{
CRC_test(Tumor, Normal)
}
\value{
test_result

pred4
}
\description{
Based on SVM method and two methylation probe data, CRC early diagnosis model was constructed
}
\examples{
CRC_test(Tumor,Normal)
CRC_predict(data1)
}
